+++ RELEASE_NOTES.rst
@@ -63,6 +63,7 @@ Maintenance
 * Add virtual axes.
 * Complete 'refineLattice()' for Core and Sample.
 * Compute B from sample lattice.
+* Control displayed precision of position tuples using 'position_display_digits'.
 * Consistent response when no forward solutions are found.
 * Engineering units throughout
     * Solver defines the units it uses.